feat: useHandleDiaryType  주석 처리 후 컴포넌트 함수 내부에 api 연결 코드 작성
