+++ src/components/OnBoardingSteps/OnBoardingStep5/OnBoardingStep5.jsx
@@ -1,13 +1,27 @@
 import * as S from "../OnBoardingStep.style";
+import { usePostDiaryType } from "../../../hooks/queries/onboarding/usePostDiaryType";
+import { useRecoilValue } from "recoil";
+import { userDiaryType } from "../../../recoil/atoms";
 import SetDiaryStyle from "../../SetDiaryStyle/SetDiaryStyle";
 import Title from "../title/Title";
 import Caption from "../title/Caption";
 import BtnComplete from "../../common/buttons/complete/BtnComplete";
-import useHandleDiaryType from '../../../hooks/diary/useHandelDiaryType';
-
 
 export default function OnBoardingStep5({onNext}){
-    const handleDiaryType = useHandleDiaryType;
+    const diaryTypeAtom = useRecoilValue(userDiaryType);
+    const { mutation } = usePostDiaryType();
+
+    const handlePostDiaryStyle= () => {
+         const body = {
+            userDiaryType : diaryTypeAtom,
+         };
+ 
+         mutation.mutate(body, {
+             onSucess: (data) => {
+                 console.log(data);
+             }
+         })
+     };
 
     return (
         <S.OnBoardingStepWrapper>
@@ -29,7 +43,7 @@ export default function OnBoardingStep5({onNext}){
             <S.OnBoardingFooter>
                 <Caption>일기 작성 페이지는 마이페이지에서 수정 가능합니다.</Caption>
                 <BtnComplete onClick={() => {
-                        handleDiaryType();
+                        handlePostDiaryStyle();
                         onNext()
                     }}
                 >
